Global refactor. MaxwellDefs and 1D are grouped and renamed to MaxwellEvolutionMethods.
