+++ src/maxwell/MaxwellEvolution1D.h
@@ -6,8 +6,7 @@
 #include "Model.h"
 #include "Sources.h"
 #include "SourcesManager.h"
-#include "MaxwellDefs.h"
-#include "MaxwellDefs1D.h"
+#include "MaxwellEvolutionMethods.h"
 
 namespace maxwell {
