+++ test/MFEMHesthavenComparison/MfemHesthaven2DTest.cpp
@@ -4,8 +4,7 @@
 #include "maxwell/Types.h"
 #include "MfemHesthavenFunctionsTest.h"
 #include "GlobalFunctions.h"
-#include "maxwell/MaxwellDefs.h"
-#include "maxwell/MaxwellDefs1D.h"
+#include "maxwell/MaxwellEvolutionMethods.h"
 
 #include "AnalyticalFunctions2D.h"
 #include "SourceFixtures.h"
